nop task fix

+++ BE/src/main/java/com/example/be/repository/IStageRepository.java
@@ -3,5 +3,5 @@ package com.example.be.repository;
 import com.example.be.model.Stage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IStageRepository extends JpaRepository<Stage, Long> {
+public interface IStageRepository extends JpaRepository<Stage, Integer> {
 }
